+++ src-tauri/src/analysis_cache/mod.rs
@@ -1,0 +1,5 @@
+mod compute;
+mod store;
+
+pub use compute::{recommended_gain_for_target, seed_from_bytes_execute, SeedFromBytesOutcome};
+pub use store::{AnalysisCache, TrackKey};
